requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
attrs==25.4.0
blake3==1.0.7
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorful==0.5.7
compressed-tensors==0.8.1
datasets==4.2.0
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
einops==0.8.1
fastapi==0.119.0
filelock==3.20.0
fire==0.7.1
flash_attn==2.8.3
Flask==3.1.2
frozenlist==1.8.0
fsspec==2024.6.1
gguf==0.10.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
lm-format-enforcer==0.10.12
MarkupSafe==2.1.5
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post6
pillow==11.0.0
platformdirs==4.5.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic_core==2.41.1
pyngrok==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.50.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.1
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.0
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
virtualenv==20.35.3
vllm==0.6.6.post1
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xformers==0.0.28.post3
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
